% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromhere-package.R
\docType{package}
\name{fromhere-package}
\alias{fromhere}
\alias{fromhere-package}
\title{fromhere: File Paths Relative to Project Roots}
\description{
Provides a set of helper functions for constructing file paths relative to the root of various types of projects, such as R packages, Git repositories, and more. File paths are specified with function arguments, or `$` to navigate into folders to specific files supported by autocompletion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mitchelloharawild/fromhere}
  \item Report bugs at \url{https://github.com/mitchelloharawild/fromhere/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
